+++ Changelog.md
@@ -5,6 +5,7 @@ Lilu Changelog
 - Ignored disabled kexts earlier for speed reasons
 - Added High Sierra to the list of compatible OS
 - Added arrsize API
+- Made patch count warning only show in debug mode
 
 #### v1.1.5
 - Increased executable memory buffer to page size